Add multi-route React app with Victim and NDRF UIs
